// packages/shared/src/schemas/content.d.ts
export declare const productSchema: z.ZodObject<{
    name_uz: z.ZodString;
    name_ru: z.ZodString;
    slug: z.ZodString;
    description_uz: z.ZodOptional<z.ZodString>;
    description_ru: z.ZodOptional<z.ZodString>;
    price: z.ZodNullable<z.ZodOptional<z.ZodNumber>>;
    stock: z.ZodNullable<z.ZodOptional<z.ZodNumber>>;
    brandId: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    categoryId: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    specsText: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    galleryIds: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    audience: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    formFactors: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    signalProcessing: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    powerLevel: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    hearingLossLevels: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    smartphoneCompatibility: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    tinnitusSupport: z.ZodNullable<z.ZodOptional<z.ZodBoolean>>;
    paymentOptions: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    availabilityStatus: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    intro_uz: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    intro_ru: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    features_uz: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    features_ru: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    benefits_uz: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    benefits_ru: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    tech_uz: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    tech_ru: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    fittingRange_uz: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    fittingRange_ru: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    regulatoryNote_uz: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    regulatoryNote_ru: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    galleryUrls: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    relatedProductIds: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    usefulArticleSlugs: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    status: z.ZodDefault<z.ZodEnum<["published", "draft", "archived"]>>;
}, "strip", z.ZodTypeAny, {
    status?: "published" | "draft" | "archived";
    slug?: string;
    name_uz?: string;
    name_ru?: string;
    description_uz?: string;
    description_ru?: string;
    price?: number;
    stock?: number;
    brandId?: string;
    categoryId?: string;
    specsText?: string;
    galleryIds?: string[];
    audience?: string[];
    formFactors?: string[];
    signalProcessing?: string;
    powerLevel?: string;
    hearingLossLevels?: string[];
    smartphoneCompatibility?: string[];
    tinnitusSupport?: boolean;
    paymentOptions?: string[];
    availabilityStatus?: string;
    intro_uz?: string;
    intro_ru?: string;
    features_uz?: string[];
    features_ru?: string[];
    benefits_uz?: string[];
    benefits_ru?: string[];
    tech_uz?: string;
    tech_ru?: string;
    fittingRange_uz?: string;
    fittingRange_ru?: string;
    regulatoryNote_uz?: string;
    regulatoryNote_ru?: string;
    galleryUrls?: string[];
    relatedProductIds?: string[];
    usefulArticleSlugs?: string[];
}, {
    status?: "published" | "draft" | "archived";
    slug?: string;
    name_uz?: string;
    name_ru?: string;
    description_uz?: string;
    description_ru?: string;
    price?: number;
    stock?: number;
    brandId?: string;
    categoryId?: string;
    specsText?: string;
    galleryIds?: string[];
    audience?: string[];
    formFactors?: string[];
    signalProcessing?: string;
    powerLevel?: string;
    hearingLossLevels?: string[];
    smartphoneCompatibility?: string[];
    tinnitusSupport?: boolean;
    paymentOptions?: string[];
    availabilityStatus?: string;
    intro_uz?: string;
    intro_ru?: string;
    features_uz?: string[];
    features_ru?: string[];
    benefits_uz?: string[];
    benefits_ru?: string[];
    tech_uz?: string;
    tech_ru?: string;
    fittingRange_uz?: string;
    fittingRange_ru?: string;
    regulatoryNote_uz?: string;
    regulatoryNote_ru?: string;
    galleryUrls?: string[];
    relatedProductIds?: string[];
    usefulArticleSlugs?: string[];
}>;